Use int* instead of size_t* to store sparsity

+++ uno/ingredients/hessian_models/IdentityHessian.cpp
@@ -23,11 +23,11 @@ namespace uno {
       return model.number_variables;
    }
 
-   void IdentityHessian::compute_sparsity(const Model& model, size_t* row_indices, size_t* column_indices, size_t solver_indexing) const {
+   void IdentityHessian::compute_sparsity(const Model& model, int* row_indices, int* column_indices, int solver_indexing) const {
       // diagonal structure
       for (size_t variable_index: Range(model.number_variables)) {
-         row_indices[variable_index] = variable_index + solver_indexing;
-         column_indices[variable_index] = variable_index + solver_indexing;
+         row_indices[variable_index] = static_cast<int>(variable_index) + solver_indexing;
+         column_indices[variable_index] = static_cast<int>(variable_index) + solver_indexing;
       }
    }
