+++ uno/optimization/OptimizationProblem.cpp
@@ -49,13 +49,13 @@ namespace uno {
       return hessian_model.number_nonzeros(this->model);
    }
 
-   void OptimizationProblem::compute_constraint_jacobian_sparsity(size_t* row_indices, size_t* column_indices, size_t solver_indexing,
+   void OptimizationProblem::compute_constraint_jacobian_sparsity(int* row_indices, int* column_indices, int solver_indexing,
          MatrixOrder matrix_order) const {
       this->model.compute_constraint_jacobian_sparsity(row_indices, column_indices, solver_indexing, matrix_order);
    }
 
-   void OptimizationProblem::compute_hessian_sparsity(const HessianModel& hessian_model, size_t* row_indices,
-         size_t* column_indices, size_t solver_indexing) const {
+   void OptimizationProblem::compute_hessian_sparsity(const HessianModel& hessian_model, int* row_indices,
+         int* column_indices, int solver_indexing) const {
       hessian_model.compute_sparsity(this->model, row_indices, column_indices, solver_indexing);
    }
